Add/delete places and questions functions and buttons handling

+++ client/src/app/places/places.component.ts
@@ -32,15 +32,41 @@ export class PlacesComponent implements OnInit {
 
 
   public deletePlaceById(place : Place){ 
-    this.apiService.deletePlaceById(place.id).subscribe(
-      res => {
-        let indexOfPlace = this.places.indexOf(place);
-        this.places.splice(indexOfPlace,1);
-      },
+    if (confirm("Are you sure, you want to delete "+place.name+" from table places?"))
+      this.apiService.deletePlaceById(place.id).subscribe(
+        res => {
+          let indexOfPlace = this.places.indexOf(place);
+          this.places.splice(indexOfPlace,1);
+        },
+        err => {
+          console.log(err);
+        }
+      );
+  }
+
+
+  public addNewPlace(){
+    let newPlace : Place = {
+      description : "Place description",
+      name : "Place name",
+      points : 0,
+      visits : 0,
+      rating : 0.0,
+      id : null, 
+      latitude : null, 
+      year : null,
+      longitude : null
+    }
+    this.places.push(newPlace);
+  }
+
+
+  public modifyPlace(place : Place){
+    this.apiService.savePlaceChanges(place).subscribe(
+      res => {},
       err => {
         console.log(err);
       }
-    );
+    )
   }
-
 }
