+++ client/src/app/quiz/quiz.component.ts
@@ -1,6 +1,5 @@
 import { Component, OnInit } from '@angular/core';
 import { Question } from './model/question';
-import { HttpClient } from '@angular/common/http';
 import { ApiService } from '../api/api-service.service';
 
 @Component({
@@ -31,14 +30,46 @@ export class QuizComponent implements OnInit {
 
 
   public deleteQuestion(question : Question){
-    this.apiService.deleteQuizQuestionById(question.id).subscribe(
-      res => {
-        let indexOfQuestion = this.questions.indexOf(question);
-        this.questions.splice(indexOfQuestion,1);
-      },
-      err => {
-        alert("Error...");
+    
+    if (confirm("Are you sure, you want to delete question #"+question.id+" from table questions?")){
+     if (question.id === null) {
+      let indexOfQuestion = this.questions.indexOf(question);
+      this.questions.splice(indexOfQuestion,1);
+     }
+     else{
+      this.apiService.deleteQuizQuestionById(question.id).subscribe(
+        res => {
+          let indexOfQuestion = this.questions.indexOf(question);
+          this.questions.splice(indexOfQuestion,1);
+        },
+        err => {
+          alert("Error...");
+        }
+      );
+     }
+    }
+  }
+
+
+  public addNewQuestion(){
+    let newQuestion : Question = {
+        a: null,
+        b: null,
+        c: null,
+        d: null,
+        correctAnswer:  null,
+        id:  null,
+        question:  null,
       }
-  );
+    this.questions.push(newQuestion);
+  }
+
+
+  public modifyQuestion(question : Question){
+    this.apiService.saveQuestionChanges(question).subscribe(
+      res => {},
+      err => {  console.log(err);}
+    )
+
   }
 }
